connecting things
Updated the legend and render_ui functions with badges instead of boxes, added waiter spinners to render_UI, added isolates around reactives that we don't want to cause an update. Commented out model 2 stuff because nothing was rendering (because there is no predict function to supply output to render_UI).